refactor: CardListPage .jsx -> .tsx
