+++ src/components/pages/CardListPage/CardListItem/styled.tsx
