+++ src/components/pages/CardListPage/styled.tsx
